MonProjet/BACKEND_API_REST_SELECTED_FILES_v1/BACKEND_API_REST_SELECTED_FILES/01_primary_obsidia_lab_trad/canonical_api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pynacl>=1.5.0
